gamification/achievement-api/requirements.txt
flask==2.0.1
azure-data-tables==12.4.0
gunicorn==20.1.0
